+++ src/thread_pool.cpp
@@ -11,3 +11,4 @@ join_threads::~join_threads()
     threads[i].join();
   }
 }
+